+++ NEWS.md
@@ -5,7 +5,9 @@
 * Add tests to geometric filters  
 * Minimum sf version increased   
 * Minimum dbplyr version now 2.2.0  
-* Migration of geoserver platform now supported as the default
+* Migration of geoserver platform now supported as the default  
+* By default in `listLayers()` pulls down abstracts and metadataID, argument provided to only pull down titles    
+* Suppress message when first using filter
 
 # VicmapR 0.1.9  
 * Addition of tutorial for new users, written by Rachel Swain (Vicmap for Beginners) (PR#36, #41)